refactor: move shared constants/enums out of message impl files
