+++ packages/core/src/capabilities/RFRegion.ts
@@ -1,0 +1,15 @@
+export enum RFRegion {
+	"Europe" = 0x00,
+	"USA" = 0x01,
+	"Australia/New Zealand" = 0x02,
+	"Hong Kong" = 0x03,
+	"India" = 0x05,
+	"Israel" = 0x06,
+	"Russia" = 0x07,
+	"China" = 0x08,
+	"USA (Long Range)" = 0x09,
+	"Japan" = 0x20,
+	"Korea" = 0x21,
+	"Unknown" = 0xfe,
+	"Default (EU)" = 0xff,
+}
