--- packages/zwave-js/src/lib/controller/ZWaveLibraryTypes.ts
@@ -1,14 +1,0 @@
-export enum ZWaveLibraryTypes {
-	"Unknown",
-	"Static Controller",
-	"Controller",
-	"Enhanced Slave",
-	"Slave",
-	"Installer",
-	"Routing Slave",
-	"Bridge Controller",
-	"Device under Test",
-	"N/A",
-	"AV Remote",
-	"AV Device",
-}
