+++ packages/zwave-js/src/lib/serialapi/_Types.ts
@@ -1,0 +1,19 @@
+export enum ZWaveLibraryTypes {
+	"Unknown",
+	"Static Controller",
+	"Controller",
+	"Enhanced Slave",
+	"Slave",
+	"Installer",
+	"Routing Slave",
+	"Bridge Controller",
+	"Device under Test",
+	"N/A",
+	"AV Remote",
+	"AV Device",
+}
+
+export enum NodeIDType {
+	Short = 0x01,
+	Long = 0x02,
+}
