+++ packages/zwave-js/src/lib/serialapi/capability/SerialAPISetupMessages.ts
@@ -2,6 +2,7 @@ import {
 	MessageOrCCLogEntry,
 	MessageRecord,
 	parseBitMask,
+	RFRegion,
 	validatePayload,
 	ZWaveError,
 	ZWaveErrorCodes,
@@ -24,6 +25,7 @@ import {
 	priority,
 } from "../../message/Message";
 import type { SuccessIndicator } from "../../message/SuccessIndicator";
+import { NodeIDType } from "../_Types";
 
 export enum SerialAPISetupCommand {
 	Unsupported = 0x00,
@@ -36,27 +38,6 @@ export enum SerialAPISetupCommand {
 	GetRFRegion = 0x20,
 	SetRFRegion = 0x40,
 	SetNodeIDType = 0x80,
-}
-
-export enum RFRegion {
-	"Europe" = 0x00,
-	"USA" = 0x01,
-	"Australia/New Zealand" = 0x02,
-	"Hong Kong" = 0x03,
-	"India" = 0x05,
-	"Israel" = 0x06,
-	"Russia" = 0x07,
-	"China" = 0x08,
-	"USA (Long Range)" = 0x09,
-	"Japan" = 0x20,
-	"Korea" = 0x21,
-	"Unknown" = 0xfe,
-	"Default (EU)" = 0xff,
-}
-
-export enum NodeIDType {
-	Short = 0x01,
-	Long = 0x02,
 }
 
 function testResponseForSerialAPISetupRequest(
