+++ packages/zwave-js/src/lib/serialapi/network-mgmt/GetRoutingInfoMessages.ts
@@ -1,11 +1,12 @@
 import { MessageOrCCLogEntry, NUM_NODEMASK_BYTES } from "@zwave-js/core";
 import type { JSONObject } from "@zwave-js/shared";
-import type { Driver } from "../driver/Driver";
+import { parseNodeBitMask } from "../../controller/NodeBitMask";
+import type { Driver } from "../../driver/Driver";
 import {
 	FunctionType,
 	MessagePriority,
 	MessageType,
-} from "../message/Constants";
+} from "../../message/Constants";
 import {
 	expectedResponse,
 	Message,
@@ -13,8 +14,7 @@ import {
 	MessageDeserializationOptions,
 	messageTypes,
 	priority,
-} from "../message/Message";
-import { parseNodeBitMask } from "./NodeBitMask";
+} from "../../message/Message";
 
 interface GetRoutingInfoRequestOptions extends MessageBaseOptions {
 	nodeId: number;
